refactor(tools): split ExecutionsTool helpers into focused modules
Extract conversation formatting, run-directory traversal, config field
filtering, and wait coordination from the 1094-line ExecutionsTool into
src/tools/executions/; the tool class becomes a thin path router with an
unchanged public surface.
